fixup! [IMP] web_chatter_position: improve form and chatter layout styles
* Fix missing classes when chatter is forced aside, leading to a much smaller width
of the chatter.
* Fix undesirable scrolling of the sheet instead of scrolling of the overflowing
elements of the sheet when chatter is aligned to the side.
* Remove ineffective reference to .o-mail-Form-chatter aimed to make chatter go
full width as well, which is a matter of preference. The main purpose of including
web_sheet_full_width's css is to make the *sheet* use the space on the sides, not so
much the chatter.
* Put new form view css into existing css file.